Change proxy auth setup in launcher to always happen
We need to do it globally on startup, not only when checking
for a new version. That way it also works for "activator new".
Once we do it, we should not need to supply a Proxy to
URL.openConnection - if we did need a custom openConnection,
we'd have to go fix a whole bunch of other openConnection cases
too, such as in sbt-io.